+++ witdim_frontend-web/src/App.tsx
@@ -1,13 +1,11 @@
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 
-import Navbar from "./components/Navbar/Navbar";
-import Content from "./components/Content/Content";
+import Content from "./containers/Content/Content";
 
 const App = () => {
   return (
     <Router>
-      <Navbar />
-      {/* Placeholder Title LoL */}
+      {/* Placeholder logo LoL */}
       <h1 style={{ textAlign: "center" }}>WitDim</h1>
       <Content />
     </Router>