+++ witdim_frontend-web/CHANGELOG.md
@@ -7,6 +7,11 @@ Addition of markdown posts will not be versioned or require a changelog update.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - 2021-09-28
+
+- Moved away from navbar and added utilbar under logo-placeholder
+- UI still heavily in development
+
 ## [0.0.3] - 2021-09-28
 
 - Made huge improvements with search